+++ src/commands/misc/qrcode.js
@@ -5,7 +5,8 @@
  */
 module.exports = {
     name: "qrcode",
-    description: "Generate/Read a qrcode",
+    aliases: ["qr"],
+    //description: "Generate/Read a qrcode",
     usage: "[subcommand] [data / image_url]",
     subcommand: true,
     subcommands: [